Major rework,
Major redesign for both GUI and core functionality,
added max volume option
added pop-up for volume controls
added new theme files for pre API:21 and Post API:21
Selected target API as 22 to prevent issues with runtime permissions.

+++ app/src/main/java/xml_mike/radioalarm/models/Alarm.java
@@ -65,7 +65,7 @@ public interface Alarm extends Parcelable{
 
     void setupAlarmData(final Context context, final ThreadedMediaPlayer mediaPlayer) throws java.io.IOException;
 
-    float getMaxVolume();
+    int getMaxVolume();
 
-    void setMaxVolume(float volume);
+    void setMaxVolume(int volume);
 }
